[WIP] Experiment with grouping loop invariant code.
NOT FOR SUBMISSION.

This patch is work in progress. The purpose is to experiment with
a similar approach to expression DAG balancing and visit inner loop
basic blocks while identifying binary trees of associative and
commutative operations and using swaps between instruction inputs
to move loop dependant code to the bottom of the tree, thus providing
the LICM optimization with more opportunities to hoist code out of
the loop.

Test: test_art_target.sh, test_art_host.sh
Change-Id: I56c2b4249fb1fa7d38cc3370c2d2bee65b2c0654