Ensure proxy methods are reported by instrumentation
Previously we would not report proxy methods to instrumentation in all
cases. Specifically if the instrumentation trampolines were not
installed we would never report proxy method entry or exit. This fixes
that and ensures that proxy methods are always reported to
instrumentation regardless of what instrumentation method is being
used. It also ensures that these methods will only ever be reported
once per entry.

Test: ./test.py --host
Test: ./test/testrunner/testrunner.py \
         --host \
         --run-test-option='--with-agent libtifast.so=MethodEntry,MethodExit'

Change-Id: I78f1d4b3b3f242e8915989ab26f7c204baab08e5