Update patch set 6
Patch Set 6: Presubmit-Verified+1

=== Finished presubmit run: L08000000182014471 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/704664?revision=6&workplanId=L08000000182014471




TreeHugger finished with: 18 passed (2 new).

























WARNING: The tip-of-tree(4837754) for git_master-art-host/art-test may be broken.
The presubmit build passes with the last-known-good-build(4837017), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



WARNING: The tip-of-tree(4837754) for git_master-art-host/art-test-javac may be broken.
The presubmit build passes with the last-known-good-build(4837017), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



This is an awesome step forward. Great job!

Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 6
Tag: autogenerated:TreeHugger:Verify-Pass
Reviewer: Gerrit User 1062513 <1062513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=+1