Update patch set 5
Patch Set 5:

(1 comment)

> Patch Set 5:
> 
> Here's another suggestion to avoid the new data structure:
> - Put the JIT entrypoint in the profiling info
> - If that messes up with the code collection, dynamically change garbage_collect_code_ to false
> 
> Would that work?

That would probably work. I think turning off jit-gc is something we really want to avoid though (in addition to the reasons given on the bug it is likely to hide issues).

I think we might use *more* memory doing it that way though since we'll need to add kRuntimePointerSize bytes for every active ProfilingInfo anywhere. Using the map here we need 2*kRuntimePointerSize bytes for every compiled method. AFAIK we compile nowhere near 50% of executed methods so the map should be more efficient I think.

I tried to re-use the already existing SavedEntrypoint field but I was never able to get that to come even close to working. That field is so bound up in the jit GC that all my attempts to change it bought me was a ton of random crashes when I turned on tracing.

Patch-set: 5
Reviewer: Gerrit User 1047769 <1047769@85c56323-6fa9-3386-8a01-6480fb634889>