Update patch set 2
Patch Set 2:

=== Finished presubmit run: L08900000182029419 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/703759?revision=2&workplanId=L08900000182029419




NOTE: This comment applies to an older patchset.

TreeHugger finished with: 16 passed.























WARNING: The tip-of-tree(4837754) for git_master-art-host/art-test may be broken.
The presubmit build passes with the last-known-good-build(4837017), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



This is some super-exciting work! Well done!

Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 2
Tag: autogenerated:TreeHugger:Verify-Stale-Pass