Update patch set 7
Patch Set 7:

=== Finished presubmit run: L26000000182024167 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/703781?revision=7&workplanId=L26000000182024167




NOTE: Since this build started, a new topic was set or a new patch was uploaded.

TreeHugger finished with: 18 passed.

























WARNING: The tip-of-tree(4837754) for git_master-art-host/art-test may be broken.
The presubmit build passes with the last-known-good-build(4837017), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



WARNING: The tip-of-tree(4837754) for git_master-art-host/art-test-javac may be broken.
The presubmit build passes with the last-known-good-build(4837017), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



I am TreeHugger, and I have hugged this change!

Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 7
Tag: autogenerated:TreeHugger:Verify-Stale-Pass