Update patch set 1
Patch Set 1:

(1 comment)

This code would be slightly easier to read if the Monitor constructors had a comment that hashcode can be zero. I guess that can be deduced from the hash_code_ comment, I still found it odd that you must pass it in, but it may be zero.

I'm neutral on the other discussion. A comment saying that this is unnecessary but benign if the CAS fails might help.

Patch-set: 1
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>