Update patch set 18
Patch Set 18:

so it seems what's going on is the jit will unconditionally rewrite new-instance String to a call to newEmptyString. Normally this is removed because it is an intrinsic and replaced with a load of an empty string which is then removed since it's not actually used. By turning off the intrinsics we instead end up with the invoke still there and we cannot remove it since invokes are visible as far as the simplifier is concerned.

It might be good to add some pass to the instruction simplifier to have it try to remove the HNewInstance directly instead of rewriting it to an HInvoke and expecting the combination of DCE and intrinsics to take care of it but for now we will just accept the jit-compiled trace as legitimate.

Patch-set: 18
Reviewer: Gerrit User 1047769 <1047769@85c56323-6fa9-3386-8a01-6480fb634889>