Update patch set 21
Patch Set 21:

> Patch Set 21:
> 
> > Patch Set 18:
> > 
> > so it seems what's going on is the jit will unconditionally rewrite new-instance String to a call to newEmptyString. Normally this is removed because it is an intrinsic and replaced with a load of an empty string which is then removed since it's not actually used. By turning off the intrinsics we instead end up with the invoke still there and we cannot remove it since invokes are visible as far as the simplifier is concerned.
> > 
> > It might be good to add some pass to the instruction simplifier to have it try to remove the HNewInstance directly instead of rewriting it to an HInvoke and expecting the combination of DCE and intrinsics to take care of it but for now we will just accept the jit-compiled trace as legitimate.
> 
> Don't we want String.<init> to be reported instead? Not reporting anything or reporting eg newEmptyString looks as bad to me.

I don't think so. Tracing like this is meant to give a programmer insight into the code that is actually running. In this case that is a newEmptyString followed by a StringBuilder.toString. While I don't think we should be calling the newEmptyString I don't think we should lie about doing it.

Also we would need to lie about things like stack traces to be consistent if we wanted to do this.

Patch-set: 21