Always perform SSA removal of new-instance string
We were skipping doing SSA pruning of 'new-instance string' due to a
desire to maintain compat with the interpreter in debuggable mode.
Unfortunately the interpreters behavior has changed and it will no
longer allocate new strings on new-instance String. This could lead to
visible differences when tracing jit-compiled and interpreted code.

Needs next CL for test to work.

Bug: 110884646
Test: ./test/testrunner/testrunner.py --host --runtime-option=-Xjitthreshold:0 --jit
Change-Id: I7b4b084bcf7dd9a23485c0e3cd2cd04a04b43d3d