Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> What problem did you find? Given the caching is based on an index locally, and a name globally, I don't think caching is an issue.

My mistake, the cache is OK int this func. 
It comes from I add a func named 'VeriClass* GetVeriClass(const std::string& name)', and cannot cache value in this func.

Otherwise, I found veridex is miss on a static link of func 'TextView::notifyViewAccessibilityStateChangedIfNeeded'. This is case by the TextView does not implement notifyViewAccessibilityStateChangedIfNeeded, but inherits from the parent class. When a class inherits from TextView, and call the super's notifyViewAccessibilityStateChangedIfNeeded by static link, the HiddenApiFinder::CheckMethod will use "TextView::notifyViewAccessibilityStateChangedIfNeeded" to find in the HiddenAPIList, and mismatched.

The right thing to do is to trace the method who impl it, and get the 'View::notifyViewAccessibilityStateChangedIfNeeded'.

The same problem is in 'HiddenApiFinder::CheckField'.

To do this, you also need to replace system-dex with full system-stubs.

Patch-set: 1
Reviewer: Gerrit User 1334788 <1334788@85c56323-6fa9-3386-8a01-6480fb634889>