Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Ah I think now I remember. It's not about calling newEntryString or something else, it's more about having the string being allocated. In non debuggable the compiler doesn't allocate. But the interpreter always allocates.
> 
> David, do you remember why we wanted to preserve consistency with the interpreter in debuggable mode for string.<init>?

Yeah, it has to do with the allocation and observing the state of registers after NEW_INSTANCE of String. The interpreter always allocates and uses that reference value to find all aliases it needs to replace with the result of StringFactory. If you enable the optimization, the observable result of compiled NEW_INSTANCE will be null, which could potentially break some assumptions people make about execution, and it's a discrepancy between the interpreter and the compiler.

So given that the interpreter needs that fake alloc, I opted to do the same in the compiler.

Patch-set: 3