Fix longstanding bug around implicit NPEs and GC.
The TODO has been there since M (so forever :)):
https://android-review.googlesource.com/c/platform/art/+/122794/13//COMMIT_MSG#13

We hardly see the issue in our tests as we need to have:
1) A GC happening while creating the NPE object.
2) ParallelMoves between the NullCheck and implicit null check operation
   that moves references.

The CL moves the information that an instruction should do
an implicit null check as a field to HInstruction instead of having
to compute it during location building and code generation. Also,
we put the environment of the null check on the instruction that will
do the implicit null check (when it doesn't have an environment
already). This means we can also just remove the HNullCheck from
the graph before code generation.

Test: test.py --gcstress
Change-Id: I3ecea4fe0d7e483e93db83281ca10db47da228c5