Make GetLocalVariables not pause jit.
We were pausing the jit since we could end up confused if a jit method
finished being compiled when the method is deoptimized. We made it so
the jit could determine when jit compilation failed during the final
stages of installing the jit compiled code. This allows us to no
longer have to pause jit during GetLocalVariables.

Test: ./test.py --host -j50
Test: ./art/tools/run-libjdwp-tests.sh --mode=host
Test: see b/111131691

Bug: 111131691
Change-Id: If681b52a2dbabc118a252436e5083f99e79b3ddd