Update patch set 3
Patch Set 3:

=== Finished presubmit run: L89300000190202642 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/715725?revision=3&workplanId=L89300000190202642




NOTE: Since this build started, a new topic was set or a new patch was uploaded.

TreeHugger finished with: 20 passed.





WARNING: The tip-of-tree(4889925) for git_master/cf_x86_phone-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4889631), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




WARNING: The tip-of-tree(4889925) for git_master/errorprone may be broken.
The presubmit build passes with the last-known-good-build(4889631), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



WARNING: The tip-of-tree(4889925) for git_master/gce_x86_phone-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4889631), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.










WARNING: The tip-of-tree(4889925) for git_master/marlin-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4889631), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




WARNING: The tip-of-tree(4889925) for git_master/sailfish-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4889631), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



WARNING: The tip-of-tree(4889925) for git_master/taimen-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4889631), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




WARNING: The tip-of-tree(4889925) for git_master/walleye-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4889631), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.






Impressive, young Jedi.

Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 3
Tag: autogenerated:TreeHugger:Verify-Stale-Pass