Update patch set 1
Patch Set 1:

(3 comments)

> Patch Set 1:
> 
> (4 comments)
> 
> I know why you're doing it, but can't we just rely on thread_local for some other things? :-)

This seemed like as good a reason as any to get the cleanup so the jvmti thread-locals stuff will actually be deallocated. We could just use thread_local for the backtrace_helper and leave this as is if you want.

Patch-set: 1
Reviewer: Gerrit User 1047769 <1047769@85c56323-6fa9-3386-8a01-6480fb634889>