Pause, don't stop, the jit
There were and have been several deadlocks between openjdkjvmti and
libjdwp mostly caused by the openjdkjvmti plugin stopping the jit while
it does something. If the jit thread is paused by some agent the
plugin will wait forever for the jit thread to finish and deadlock.
Since generally we don't need the jit thread to stop but just to not
commit any inprogress work (and not pick up any new work) we added
support for pausing the jit and resuming it while discarding
in-progress work. While the jit is paused no new compile jobs will be
started. In progress jobs will continue but will suspend themselves
before committing their compiled code. If the in-progress jobs are
canceled the results of the jit compile will be thrown out.

Test: ./test/testrunner/testrunner.py --host
Test: ./art/tools/run-libjdwp-tests.sh --mode=host --debug
Bug: 111394423
Change-Id: Iacaf289095e3663deff4ebcffb4deeba09c794eb