Add some helpful features to libjdwp test runners
Add --verbose support to run-libjdwp-tests.sh. This makes it use the
libjdwp logging on the debuggee with by default log level 0xC0. This
can be changed using --verbose-level and --verbose-all. If one wants
to log to somewhere other than the debuggee's stderr use
--verbose-file.

Add an --instant-jit flag to both runners. This forces the debuggee to
run with -Xjitthreshold:0.

Test: manual
Change-Id: Iff9884b0cfbc5d32ff40c5c2d8f48b7876944509