Update patch set 4
Patch Set 4:

Ok. So it seems that when called within the entry/exit hooks the stack-walker cannot see the frames being entered and exited. This is a problem for JVMTI since it meant that things like GetLocalInstance etc don't work.

For now I'm going to just put transitioning jvmti to the trampolines on ice but still try to get in the CommitCodeInternal cleanup and a jit-pause system in since both of those are independently useful and can fix some (small) deadlocks.

Patch-set: 4
Reviewer: Gerrit User 1047769 <1047769@85c56323-6fa9-3386-8a01-6480fb634889>