Update patch set 1
Patch Set 1: Code-Review+1

(4 comments)

Unfortunately, this looks necessary and correct to me. I say "unfortunately" because it's likely to have a performance impact. Particularly in the case of Mutex, this code urgently needs to be revisited; the contention count should clearly be part of the state.

Filed b/111835365 to track the resulting performance issue.

Assuming you are OK with my changes, please change the commit message as indicated. Otherwise LGTM. I'll change my score to a +2 once we agree on the commit message.

Patch-set: 1
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1