Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> It looks like the inc/dec operations on num_contenders_ in ExclusiveLock could be acq_rel rather than cst, then ExclusiveUnlock should use load(acq).

I don't think that suffices. The problem is with the reordering of the state store in the releasing thread and the num_contenders load. The only way to prevent that reordering is to make those seq_cst or insert a seq_cst fence between them. I think this is the simplest correct fix, as is. I'd like to see the correctness issue fixed, and then address performance.

It seems easy, but still significantly more intrusive, to fix this in Mutex by combining state and num_contenders. I'll probably do that shortly after this goes in. But ReaderWriterMutex is nontrivial on 32-bit hardware.

I'm somewhat embarrassed that I didn't catch this earlier, but also surprised that this has actually been observed. The failure scenario seems to require a rather long window during which the reordering is visible. I'm concerned that we will see more of this on newer SOCs.

Patch-set: 1
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>