Update patch set 2
Patch Set 2:

(4 comments)

I don't know what the status of this is.

It seems to me that the correct solution here is to define a struct StateAndFlags, and to use an atomic<StateAndFlags>, dropping the union completely. Did Orion already try that?

Patch-set: 2
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>