Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > > > (Why isn't PGO finding this for us?)
> > > 
> > > Did you investigate in AOSP?  Currently we have PGO on in the internal P release branch.  Can you check if the functions were handled correctly in there?
> > 
> > Numbers in commit message are from AOSP, yes. For internal master, I get 4.79B -> 4.25B. In the release branch, 5B -> 4.5B.
> 
> (Note: it's all host)

PGO is not enabled for host (the reason was just to avoid adding untested/unused features that bitrot).  If it makes sense for ART, we can revisit support for the host, and also continuous updates to profiles in both AOSP and master.  We can move discussion to email if this feature will be useful.

Patch-set: 1
Reviewer: Gerrit User 1058860 <1058860@85c56323-6fa9-3386-8a01-6480fb634889>