Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Could you comment on how the boot classpath is picked up

Like other tools the boot classpath is a parameter to the invocation (usage printing is in the base class). So something like

  art_verifier --boot-image=... --instruction-set=arm64

> and how to add dex files in the classpath?

That is not implemented, yet. I aim to eventually support dex2oat-style passing of the CLC (but plan to talk to Calin about it). However, right now the tool is decent enough to verify dex files without dependency, and profile verification.

Patch-set: 2