Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> After our offline discussion, I've had a follow-up discussion with Orion about an idea we had in common.
> 
> For a simpler solution, we should do this instead:
> - Have a timestamp (that eg starts at 0), which gets updated when jvmti doesn't want JIT threads to commit code.
> - A JIT compilation thread fetches and records that timestamp when starting compilation.
> - At the point of commit, the JIT thread looks at the current timestamp and compares it to the one it fetched when starting the compilation. If they're not the same, the thread does not commit the code. If they are the same, the thread commits the code.
> 
> The timestamp value can probably be guarded by the JIT lock.

I actually considered doing it this way but there are some problems.

The way the code is structured means that putting the invalidation check anywhere other than pretty far down the stack (under JitCodeCache::CommitCode) would require a major rewrite of the jit stack. Using a list means we don't need to do the complicated task of sending the start generation all the way down to the commit-code.

Also having it with this list means it might be possible in the future to only invalidate some of the threads based on what they actually did, which we cannot do at all with a generation count.

Patch-set: 6
Reviewer: Gerrit User 1047769 <1047769@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0