Don't greylist public bridge methods.
We need to be able to apply the @UnsupportedAppUsage annotation to regular
(non-bridged) methods, but due to the SDK visibility rules, the synthetic
bridge is sometimes part of the SDK. We need to know the public API set
in order to exclude these.

While we're here, also fix a bug with volatile fields. The access bit for
bridge methods is also used for volatile fields, so ignore that bit if the
member is not a method. Add some test cases for this too.

Bug: 110868826
Test: atest class2greylisttest
Test: m
Change-Id: I2ac9dfaaf60053762f379a5dc0d81e48a40a4e57