Update patch set 5
Patch Set 5:

Note the state of this WIP CL (ooo until 8/22):

The CL needs to resolve the store in Thread::SetState() which is used to move around suspended states. The store was formerly to 16-bit state field and relaxed. It's not immediately clear that the suspend related flags (kSuspendRequest / kActiveSuspendBarrier) might not be touched in another thread (via Thread::ModifySuspendCountInternal()) whilst SetState() is running.

CL#5 has a relaxed store in SetStateUnsafe() and comments on why this should be sufficient.

Patch-set: 5
Reviewer: Gerrit User 1071150 <1071150@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0