Needs to notify for string empty case on ReadFileToString()
The function ReadFileToString() can result in some NE
failures with its ambiguous return values. Here is
suggesting an update to notify on the empty string case.

Bug:
Test: This issue occurred on MTBF-Monkey test, however,
the root cause can be verified on source code.
If a test is required, it can be a test case before applying
 this patch (https://android-review.googlesource.com/c/
platform/art/+/604097).
Since, it is one point of failure of this issue.

Change-Id: If20bf0e8c5fab1d0bbf0fe6d6068caa3ae8d3a41