Update patch set 1
Patch Set 1: Code-Review-1

The return value tells whether the file could be read or not. An empty file is successfully read, so true is the right value IMHO. The caller needs to be sensitive to an empty string, if necessary. (+enh for an opinion)

Patch-set: 1
CC: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1