Fix unhiding API in JVMTI redefine run-tests
When JVMTI redefines a boot classpath dex file, it creates a copy of
it, reverts quickening and removes hidden API access flags. This logic
would not, however, remove hidden API access flags when there is a vdex
bit no quickening info. This patch changes the logic to unhide API
unconditionally. Production code is not affected as not unhiding merely
trips a DCHECK.

Bug: 111694487
Test: art/test/testrunner/testrunner.py --interp-ac --target \
      -t 980-redefine-object -t 983-source-transform-verify
Change-Id: I3a2987e19faa0ddf299a1ebbd72cc975b93aeebf