Simplify and correct deopt/exception instrumentation interaction
Previously we would always send the ExceptionThrown instrumentation
event within QuickDeliverException. If we then deopted we would
manually skip over that event (and the MethodUnwind and
WatchedFramePopped) in EnterInterpreterFromDeoptimize. To simplify
this code and correct the missed events we instead only send the
ExceptionThrown event from QuickDeliverException if we have already
determined we are not going to deopt.

Bug: 113020907
Test: ./test.py --host -j50
Change-Id: Ic88c9e4ab2e133f4d890dcc29116dd28fd7f6698