WIP: Reorder vregs in shadow frame.
The vregs array is stored twice in the shadow frame,
where the copy is non-zero only for GC-able objects.
Maintaining the two arrays costs extra instructions.

Interleave the two arrays so that the copy directly
follows the main entry and thus both can be written
in a single 64-bit store instruction.

Change-Id: I3758269e39995965061b0f1d168e99e9ad5857e4