Implement IterateOverInstances JVMTI function
This JVMTI function can be useful since it allows one to easily
iterate over all instances of a class, including subclasses. This CL
also includes a new fieldnull tool that will allow us to examine what
portion of a fields values are null on a running process. This can be
used to determine if changing the layout or implementation of the
class is appropriate.

Bug: 32635074
Test: ./test.py --host
Test: ./test/run-test --debuggable \
                      --host \
                      --dev \
                      --with-agent 'libfieldnull.so=Ljava/lang/Class;.name:Ljava/lang/String;' \
                      001-HelloWorld
Test: am start-activity \
         -s \
         --attach-agent '/data/local/tmp/libfieldnull.so=ljava/lang/class;.name:ljava/lang/string;' \
         com.antonioleiva.bandhookkotlin/.ui.screens.main.mainactivity;
      kill -3 $(pidof com.antonioleiva.bandhookkotlin)

Change-Id: I31c3b54ca599964c64aa31fbf666253e9b8000c4