ART: Dual mapping of JIT code pages
Add support for a dual view of JIT code pages. One view is used for
executing code, the other view for updating it. Fallback to existing
single view when required support is missing.

Add compatability shims for membarrier() and memfd_create(). The
latter is required for creating a dual view.

Bug: 66095511
Bug: 111199492
Test: art/test.py --host
Test: art/test.py --64 (angler - single view)
Test: art/test.py --64 (other - dual view)

Change-Id: I9920534c81cbf16f375f1dd554d5952716bf3253