Adjust AssertToSpaceInvariantInNonMovingSpace for Sticky-Bit CC.
Do not rely on the mark bitmap to check the to-space invariant in
non-moving spaces during a minor (sticky-bit) CC collection when the
young-generation collector is still scanning the dirty cards, because
of a potential race between bitmap marking (in
ConcurrentCopying::MarkNonMoving) and to-space invariant checking (in
AssertToSpaceInvariantInNonMovingSpace). Check that the read barrier
state is gray instead.

Test: ART_USE_GENERATIONAL_CC=true art/test.py
Bug: 113858074
Bug: 67628039
Change-Id: Ifa13458bc2e28a2ef8ace1496f2e1e90b0a020e4