ART: Clean up verify and read barrier flags.
Remove read barrier option from Class::IsArrayClass() and
related functions. Comparison with null does not need the
read barrier and the primitive type should be the same when
read from the from-space version of a class as when read
from the to-space version.

Propagate the verify flags to a few more functions.

Test: m test-art-host-gtest
Test: testrunner.py --host --gcstress
Change-Id: I8b7a6165fa8e25efd37778e0e6ea049fc2147b24