Update patch set 2
Patch Set 2:

> Submitting as a stop-gap for the current breakages. Andreas could there be a potential cleaner follow-up?

It's not really a "stop gap."

I don't remember. It significantly depends on what the order between defaults-imported and directly defined libs is (and how duplicates are removed). The question is whether in a hierarchical defaults-dependency-graph it's possible to put base libraries at the end of the link list or not.

Patch-set: 2