Revert "Allow threads to be marked as unsuspendable by kForUserCode"
This allows any thread (including jit-threads) to be suspended by
kForUserCode and the will stop executing at the next suspend-point.

This reverts commit 53570676750d74416cecdf5a8e01f3cf9a8d4169.

Reason for revert: This marking was not sufficient to prevent
deadlocks and there is a better solution in not waiting for the jit to
finish at all. This is done by the next CL.

Test: ./test.py --host
Test: ./art/tools/run-libjdwp-tests.sh --mode=host
Test: ./art/tools/run-libjdwp-tests.sh \
        --mode=host                    \
        --variant=x64                  \
        --test org.apache.harmony.jpda.tests.jdwp.EventModifiers.InstanceOnlyModifierTest

Bug: 70838465
Bug: 111348762

Change-Id: I8314904cc35f66bdf287ac7b9ec69510310a3474