Improved tifast.so agent
Made tifast.so=log actually print out the values of all the arguments
passed to the events. This includes all integer values, thread names,
class names and files and method names and signatures. With log the agent
will now print lines like the following for each event:

  dalvikvm32 I 09-19 13:31:18 160583 160595 tifast.cc:579] Got event ExceptionCatch(jvmtiEnv*, JNIEnv*, jthread[FinalizerDaemon], jmethodID[Ljava/lang/Daemons$FinalizerDaemon;->runInternal()V (source: Daemons.java:217)], jlocation[88, hex: 0x58], jobject[type: Ljava/lang/InterruptedException; file: InterruptedException.java])

Gave ti-fast an 'all' option which will enable all events the runtime
is capable of supporting.

Test: ./test/run-test --host --with-agent libtifast.so=log,all 001-HelloWorld

Change-Id: I672b91495e41eea5af9a7746f4149ebb0383131e