Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Uhm, it's exactly the right thing to require JNI_TRUE/JNI_FALSE (knowing that they are 1/0 is looking at the header, not the spec).

It's in the spec, see Ch3 "Primitive Types"

> The following definition is provided for convenience.
> 
> #define JNI_FALSE  0 
> #define JNI_TRUE   1 

From a "correctness" point of view, it's pointless to write this kind of code:

  bool x = SomeBooleanExpression();
  jboolean jx = x ? JNI_TRUE : JNI_FALSE;

jboolean is defined as an unsigned int in the JNI spec and the C++ standard mandates that bool conversion is true -> 1, false -> 0

Note that for arbitary integer types, the JNI specification doesn't seem to define whether !=0 values are considered 'true'. But I think it would be confusing to ever get anything outside of {0,1} range even if it was permitted by an implementation.

So for example given

  int x = SomeIntExpression();
  jboolean jx = x ? JNI_TRUE : JNI_FALSE;

might be more reasonable due to JNI spec ambiguity, but then this shorter form is also spec-compliant:

  jboolean jx = x != 0;

Patch-set: 1
Reviewer: Gerrit User 1021471 <1021471@85c56323-6fa9-3386-8a01-6480fb634889>