Revert^2 "Test for jit deadlock"
This test seems to reveal some issue with gcstress that causes
shadow_frames to be cleared/corrupted. For now we will just disable
this test in gcstress configurations.

Also fixed up the test so it would actually redefine classes (a typo
caused it to always bail early).

This reverts commit 819824bbafc5bfea0d3896e1d3e1d4ed2188d09e.
This reverts commit 90f2e4e0342e6c17241420706322ebd74874f9b4.

Reason for revert: Disabled problematic config.
Test: ./test.py --host
Test: while ./test/run-test --host 1952; do; done;

Bug: 112074977
Bug: 70838465
Bug: 116517081

Change-Id: I49656c8cb47cea95309baa8cb5d35d834260487e