Prevent the JIT from loading classes in debuggable mode
In debuggable mode the timing and placement of class loads is
observable. This causes deopts to be placed in situations where a
class would be loaded by the JIT. In general this should have little
to no impact on performance since if the instructions causing the
class load haven't been hit by the time the JIT kicks in they are
already extremely cold.

TODO Actual numbers

Test: ./test.py --host
Test: ./test/testrunner/testrunner.py --with-agent=libjitloadd.so=fatal --host
Bug: TBD

Change-Id: I030bc8116345af506e83ba78427a7fbda7a7b386